Implement collisions with SAT
Separating Axis Theorem. The collision function checks all surface normals
looking for a gap between the two colliding convex polygons. If no gap is
found a vector is returned that represents the shortest distance necessary to
push the two polygons apart. This vector will be in the direction of the
surface normal that the collisions happened at.

When the ball collides, it is pushed the Minimum Translation Vector away
from the collision, and then gets its speed reflected along the surface normal.
This allows rotated paddles to have correct reflections.