Change Qt imageformat to reduce workload in qt
The previous imageformat was internally converted to this.
In debug this took a considerable amount of time. Using the 32-bit
format without Alpha specifier is many times faster. Alpha is disabled
for performance, but still supported by the pong-code. Qt expects it to
be there and disregards it, and pong doesn't use it for anything at the
moment. It can easily be turned back on if needed, simply by using it
and switching to ARGB32 format.